update: move sidebar component to ui from model
