+++ src/components/page/Resource/Resource.page.tsx
@@ -1,5 +1,5 @@
 import Resource from '@/src/components/page/Resource/Resource'
-import { SideBar } from '@/src/components/ui/SideBar'
+import { SideBar } from '@/src/components/model/sidebar'
 export const ResourcePage = () => {
   return (
     <SideBar>
